Change "y" for "e", adjusting title
Title:

If the "on" of "On Roasts" means "about", a better translation would be "Sobre".
"Rostizado" is used more as a degree of roasting than for roasting itself, that's why I suggest to use "Tueste". This new title could translate to "About the action of roasting"

Summary / description:

In Spanish we generally use "y" for "and", except when the second word begins with a vowel or h (because our h is silent).

Notes:

The languages are the other way around here. What's the convention you prefer? 
Should I also change the keywords in Spanish? (to tueste, probably?)