Optional translation
Only changed "27 Bowdoin St" to "Calle Bowdoin 27", but it doesn't need it.

Changing the address to Spanish doesn't really make it any clearer and, say, if I were to send something to that address, I would have to do it using the address in the custom of the country to where I'm sending something. Your call!