Translated to Spanish
Duplicated block, added id.

Went with 'hervidor de agua' for 'kettle'. This means 'water boiler'. Alternatively 'tetera' can be used, but in some places that also means 'baby bottle'. The first word that came to mind was actually 'cafetera' (coffee pot), but I don't understand if the product needs to have coffee to be used.