Update query parameter for enabling preview mode
I was looking into this, and saw that preview mode is not working. So I update the code to globalQueryConfig and now it is working, I am assuming enablePreviewMode is obsolete query parameter that is changed to the globalQueryConfig. If I am incorrect please let me know and you can ignore this PR.